+++ js/api/signup/register-account.mjs
