+++ js/handlers/registration/register-form-handler.mjs
@@ -1,4 +1,4 @@
-import { registerAccount } from "../../api/account/register-account.mjs";
+import { registerAccount } from "../../api/signup/register-account.mjs";
 import { validateInputs } from "../../form-validation/validatedInput.mjs";
 
 export function registerFormHandler() {
